Split things up & raname things
