+++ src/docs/components-interactive.html
@@ -1,5 +1,5 @@
 ---
-layout: items-react
+layout: items-interactive
 navigationIdx: 4
 bundle: components
 ---
